requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.0.0
langchain==0.3.25
langgraph==0.4.3
langchain-anthropic==0.3.13
pydantic==2.11.4